:recycle: :: (#127) add model.ts
따로 쓰는 상수나 타입을 model로 뺐다
