+++ zamong/src/components/DiaryWrite/component/Properties/Selecter/DreamDate/DreamDate.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import DatePicker from "../../../Picker/DatePicker/DatePicker";
 import * as S from "../../../styles";
-import TimePicker, { Time } from "../../../Picker/TimePicker/TimePicker";
+import TimePicker from "../../../Picker/TimePicker/TimePicker";
 import Selecter from "../Selecter/Selecter";
 const DreamDate = (): JSX.Element => {
   const [nowDate, setNowDate] = useState(new Date());
